Complete exercise 1
First line was simple and I see the advantage of using exertnal CSS.
I struggled on the second line as I had used and h2 tag for the line
and the <p> element wasn't responding. I removed the h2 tag after some
time and found it working now.

For <button> I assumed it would be the element name as button and used
the syntax of the inline CSS example on the curriculum. Everything looks
as it should per the completed example.