Upload completed exercise 4
Struggled a little with syntax of adding 2 seperate classes, thought
i had to use two class tags and wasnt working. I understood what neeeded
to be done with the chaining of classes and referred back to lesson for
syntax help. Learned that that adding containers even with no code
will seperate images on lines. Good to know.